+++ waltz-ng/client/system/system-admin-list.html
@@ -69,6 +69,18 @@
                             Recompute: Data Type Usages, Flow Authoritative Source Ratings
                         </p>
                     </div>
+
+                    <div>
+                        <a ui-sref="main.system.orphans">
+                            <waltz-icon class='text-muted'
+                                        fixed-width="true"
+                                        name="child"></waltz-icon>
+                            Orphan entities
+                        </a>
+                        <p class="text-muted small">
+                            View entities that have erroneously become orphaned
+                        </p>
+                    </div>
                 </div>
 
                 <hr>
